feat: rework auth, mailer, config, and security systems

+++ internal/db/migrations/006_add_jwt_versioning.up.sql
@@ -1,2 +1,2 @@
 ALTER TABLE users
-ADD COLUMN jwt_session_id INTEGER;
+ADD COLUMN jwt_session_id BIGINT;
